Done including tweaks from WCM - separation precision, enhanced LPR Also included MWCS input instances from DIMACS Note: the parser for .stp input was added in the previous commit

+++ io.cpp
@@ -411,8 +411,8 @@ void IO::save_lpr_info(double lp_bound, double lp_time)
     #endif
 }
 
-void IO::save_ip_info(long lb,
-                      long ub,
+void IO::save_ip_info(double lb,
+                      double ub,
                       double gap,
                       double time,
                       long node_count,
